+++ app/routes/api/v2/AttachmentsRoute.js
@@ -7,4 +7,5 @@ export default function addRoutes(app) {
   app.get('/attachments/my/stats', controller.myStats);
   app.post('/attachments/my/sanitize', controller.mySanitize);
   app.get('/attachments/:attId', controller.getById);
+  app.get('/attachments/:attId/:type', controller.getPreview);
 }